% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp500.R
\docType{data}
\name{sp500}
\alias{sp500}
\title{S&P 500 Index Price Data}
\format{
A matrix with rows representing minute-by-minute time points and columns representing intraday prices.
}
\source{
\url{https://finance.yahoo.com/}
}
\usage{
data(sp500)
}
\description{
This dataset contains S&P 500 index prices, given in percentages, for all full trading days
between January 3, 2017 and January 2, 2018, observed at a 1-minute frequency (390 grid points per day).
}
\keyword{datasets}
