% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spanish_elec.R
\docType{data}
\name{Spanish_elec}
\alias{Spanish_elec}
\title{Spanish electricity daily price profiles}
\format{
A matrix with rows representing hourly time points and columns representing the daily price profiles as observed functions.
}
\source{
\code{fdaACF}, \url{https://www.esios.ree.es/es/analisis/600}
}
\usage{
data(Spanish_elec)
}
\description{
This dataset contains daily electricity price profiles from the Day-Ahead Spanish Electricity Market.
The original data consists of hourly electricity spot prices, which have been transformed into daily functional profiles.
The dataset covers the period from January 1, 2014, to December 31, 2014.
}
\keyword{datasets}
